% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_compile.R
\name{compile_gb}
\alias{compile_gb}
\title{Compile GB raw data into database}
\usage{
compile_gb(con)
}
\arguments{
\item{con}{PqConnection: database connection}
}
\value{
string: message
}
\description{
Compile GB raw data into database
}
