% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corvus-package.R
\docType{package}
\name{corvus-package}
\alias{corvus}
\alias{corvus-package}
\title{corvus: corvus}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Measurement stations managing R shiny application.
}
\author{
\strong{Maintainer}: Louis Manière \email{louis.maniere@univ-tours.fr}

Authors:
\itemize{
  \item Louis Manière \email{louis.maniere@univ-tours.fr}
}

}
\keyword{internal}
