% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data.R
\name{data_edit_drift}
\alias{data_edit_drift}
\title{Create drift correction column to edit data frame.}
\usage{
data_edit_drift(timestamp, value_corr, drift_value)
}
\arguments{
\item{timestamp}{POSIXct: timestamp}

\item{value_corr}{numeric: corrected value}

\item{drift_value}{numeric: drift value}
}
\value{
numeric
}
\description{
Create drift correction column to edit data frame.
}
\examples{
data <- data.frame(timestamp = as.POSIXct(c("2021-01-01 00:00", "2021-01-02 00:00",
                                            "2021-01-03 00:00"), format = "\%Y-\%m-\%d \%H:\%M"),
                  value_corr = c(1, 2, 3)) \%>\%
                  dplyr::mutate(edit = data_edit_drift(timestamp, value_corr, 5))
}
