% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data.R
\name{data_get_corr_data}
\alias{data_get_corr_data}
\title{Get corr data from database}
\usage{
data_get_corr_data(con, station, parameter, start_date, end_date)
}
\arguments{
\item{con}{PqConnection: database connection}

\item{station}{character: station name}

\item{parameter}{character: parameter name}

\item{start_date}{character: start date in format 'YYYY-MM-DD'}

\item{end_date}{character: end date in format 'YYYY-MM-DD'}
}
\value{
data.frame
}
\description{
Get corr data from database
}
\examples{
con <- db_con()
data_get_corr_data(con, "BE", "level", "2019-01-05", "2021-12-26")
}
