% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data.R
\name{data_get_correction_period}
\alias{data_get_correction_period}
\title{Format correction period from edited values}
\usage{
data_get_correction_period(
  dataframe,
  sensor_id,
  value1,
  value2,
  author_id,
  correction_type,
  comment
)
}
\arguments{
\item{dataframe}{data.frame: data frame with ts, value columns}

\item{sensor_id}{integer: sensor id}

\item{value1}{numeric: tool parameter1}

\item{value2}{numeric: tool parameter2}

\item{author_id}{integer: author id}

\item{correction_type}{integer: correction type id}

\item{comment}{character: comment}
}
\value{
data.frame
}
\description{
Format correction period from edited values
}
