% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data.R
\name{data_get_intervention}
\alias{data_get_intervention}
\title{Get all the intervention from the database}
\usage{
data_get_intervention(con, station_id, start_date, end_date)
}
\arguments{
\item{con}{PqConnection: database connection}

\item{station_id}{integer: station id}

\item{start_date}{POSIXct: start date in format 'YYYY-MM-DD'}

\item{end_date}{POSIXct: end date in format 'YYYY-MM-DD'}
}
\value{
data.frame
}
\description{
Get all the intervention from the database
}
