% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data.R
\name{data_get_measurement}
\alias{data_get_measurement}
\title{Get measurement data from database}
\usage{
data_get_measurement(con, sensor, start_date, end_date)
}
\arguments{
\item{con}{PqConnection: database connection}

\item{sensor}{integer: sensor id}

\item{start_date}{character: start date in format 'YYYY-MM-DD'}

\item{end_date}{character: end date in format 'YYYY-MM-DD'}
}
\value{
data.frame
}
\description{
Get measurement data from database
}
\examples{
con <- db_con()
data_get_measurement(con, 2, "2019-01-05", "2021-12-26")
}
