% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data.R
\name{data_get_measurement_edit}
\alias{data_get_measurement_edit}
\title{Get measurement data with applied corrections}
\usage{
data_get_measurement_edit(measurement_raw, measurement_corr)
}
\arguments{
\item{measurement_raw}{data.frame: raw measurement data with columns ts, sensor_id, value}

\item{measurement_corr}{data.frame: correction data with columns ts, sensor_id, value}
}
\value{
data.frame with columns ts, sensor_id, value (corrected)
}
\description{
This function takes the raw measurement data and the corresponding corrections, and applies the corrections to the raw data. It performs a left join between the raw measurement data and the correction data based on the timestamp and sensor ID. The corrected value is calculated using the `coalesce` function, which returns the corrected value if it exists, or the original value if there is no correction. The resulting data frame contains the timestamp, sensor ID, and the final value after applying corrections.
}
