% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data.R
\name{data_get_missing_period}
\alias{data_get_missing_period}
\title{Get measurement missing period by interval}
\usage{
data_get_missing_period(con, sensor, start_date, end_date, interval_time)
}
\arguments{
\item{con}{PqConnection: database connection}

\item{sensor}{integer: sensor id}

\item{start_date}{POSIXct: start date in format 'YYYY-MM-DD'}

\item{end_date}{POSIXct: end date in format 'YYYY-MM-DD'}

\item{interval_time}{character: interval in format '1 day', '1 hour', '1 minute', '1 second'}
}
\value{
data.frame
}
\description{
Get measurement missing period by interval
}
\examples{
con <- db_con()
data_get_missing_data(con, 1, "2019-01-05", "2021-12-26", "15 minutes")
}
