% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data.R
\name{data_get_raw_data}
\alias{data_get_raw_data}
\title{Get raw data from database}
\usage{
data_get_raw_data(con, station, parameter)
}
\arguments{
\item{con}{PqConnection: database connection}

\item{station}{character: station name}

\item{parameter}{character: parameter name}
}
\value{
data.frame
}
\description{
Get raw data from database
}
\examples{
con <- db_con()
data_get_raw_data(con, "BE", "level")
}
