% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data.R
\name{data_get_stations}
\alias{data_get_stations}
\title{Get all the station from database}
\usage{
data_get_stations(con)
}
\arguments{
\item{con}{PqConnection: database connection}
}
\value{
data.frame
}
\description{
Get all the station from database
}
\examples{
con <- db_con()
data_get_stations(con)
}
