% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data.R
\name{data_hampel_filter}
\alias{data_hampel_filter}
\title{Filter value with Hampel filter}
\usage{
data_hampel_filter(x, k = 3)
}
\arguments{
\item{x}{POSIXct: date time date in format 'YYYY-MM-DD'}

\item{k}{number: Hampel factor threshold}
}
\value{
value
}
\description{
Filter value with Hampel filter
}
