% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data.R
\name{data_insert_offset}
\alias{data_insert_offset}
\title{Insert offset data into database}
\usage{
data_insert_offset(
  con,
  station,
  parameter,
  date_time_start,
  date_time_end,
  offset_val,
  author,
  comment
)
}
\arguments{
\item{con}{PqConnection: database connection}

\item{station}{character: station name}

\item{parameter}{character: parameter name}

\item{date_time_start}{character: start date in format 'YYYY-MM-DD HH:MM'}

\item{date_time_end}{character: end date in format 'YYYY-MM-DD HH:MM'}

\item{offset_val}{numeric: offset value}

\item{author}{character: author name}

\item{comment}{character: comment}
}
\value{
character
}
\description{
Insert offset data into database
}
\examples{
con <- db_con()
data_insert_offset(con, "be", "level",
  as.POSIXct("2021-01-01 00:00", format = "\%Y-\%m-\%d \%H:\%M"),
   as.POSIXct("2021-01-02 00:00", format = "\%Y-\%m-\%d \%H:\%M"),
   0.5, "Louis Manière", "my comment")
}
