% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data.R
\name{data_update_measurement}
\alias{data_update_measurement}
\title{Update measurement data into database}
\usage{
data_update_measurement(
  con,
  data,
  sensor,
  author,
  correction_type,
  value,
  comment
)
}
\arguments{
\item{con}{PqConnection: database connection}

\item{data}{data.frame: data frame with timestamp, value_corr and edit columns}

\item{sensor}{integer: sensor id}

\item{author}{integer: author id}

\item{correction_type}{character: correction type id}

\item{value}{numeric: offset value}

\item{comment}{character: comment}
}
\value{
character
}
\description{
Update measurement data into database
}
