% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_db.R
\name{db_get_all_stations}
\alias{db_get_all_stations}
\title{Get all the station from database}
\usage{
db_get_all_stations(con)
}
\arguments{
\item{con}{PqConnection: database connection}
}
\value{
data.frame
}
\description{
Get all the station from database
}
\examples{
con <- db_con()
db_get_all_stations(con)
}
