% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_db.R
\name{db_get_authors}
\alias{db_get_authors}
\title{authors list}
\usage{
db_get_authors(con)
}
\arguments{
\item{con}{PqConnection: database connection}
}
\value{
vector
}
\description{
This function returns a list of authors from the database.
}
