% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_db.R
\name{db_get_deleted_period_data}
\alias{db_get_deleted_period_data}
\title{Deleted period data}
\usage{
db_get_deleted_period_data(con, sensor_id, start_date, end_date)
}
\arguments{
\item{con}{PqConnection: database connection}

\item{sensor_id}{integer: sensor id}

\item{start_date}{POSIXct: start date in format 'YYYY-MM-DD'}

\item{end_date}{POSIXct: end date in format 'YYYY-MM-DD'}
}
\value{
data.frame
}
\description{
This function returns the deleted period data based on the sensor id and the date range.
}
