% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_db.R
\name{db_get_fieldwork_data}
\alias{db_get_fieldwork_data}
\title{Field data}
\usage{
db_get_fieldwork_data(con, station_id, start_date, end_date)
}
\arguments{
\item{con}{PqConnection: database connection}

\item{station_id}{integer: station id}

\item{start_date}{POSIXct: start date in format 'YYYY-MM-DD'}

\item{end_date}{POSIXct: end date in format 'YYYY-MM-DD'}
}
\value{
data.frame
}
\description{
This function returns the field data based on the station id and the date range.
}
