% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_db.R
\name{db_get_interval}
\alias{db_get_interval}
\title{Get main interval sensor measurement.}
\usage{
db_get_interval(con, sensor_id)
}
\arguments{
\item{con}{PqConnection: database connection}

\item{sensor_id}{integer: sensor id}
}
\value{
data.frame
}
\description{
Get main interval sensor measurement.
}
