% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_db.R
\name{db_get_parameter_unit}
\alias{db_get_parameter_unit}
\title{get parameter unit}
\usage{
db_get_parameter_unit(con, parameter_id)
}
\arguments{
\item{con}{PqConnection: database connection}

\item{parameter_id}{integer: parameter id}
}
\value{
character
}
\description{
get parameter unit
}
