% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_db.R
\name{db_get_parameters}
\alias{db_get_parameters}
\title{Get parameters list}
\usage{
db_get_parameters(con)
}
\arguments{
\item{con}{PqConnection: database connection}
}
\value{
vector
}
\description{
Get parameters list
}
