% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_db.R
\name{db_get_station_parameters}
\alias{db_get_station_parameters}
\title{parameters list}
\usage{
db_get_station_parameters(con, station_id)
}
\arguments{
\item{con}{PqConnection: database connection}

\item{station_id}{integer: station id}
}
\value{
vector
}
\description{
This function returns a list of parameters from the database.
}
