% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_db.R
\name{db_insert_station}
\alias{db_insert_station}
\title{Insert station into database}
\usage{
db_insert_station(db_con, station)
}
\arguments{
\item{db_con}{PqConnection: database connection}

\item{station}{list: list of station information (id, name, lat, lon)}
}
\value{
list
}
\description{
Insert station into database
}
