% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_db.R
\name{db_update_correction}
\alias{db_update_correction}
\title{Insert correction periods into correction table}
\usage{
db_update_correction(con, dataframe)
}
\arguments{
\item{con}{DBI Connection}

\item{dataframe}{data.frame}
}
\value{
integer Number of rows inserted
}
\description{
Insert correction periods into correction table
}
