% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data.R
\name{get_min_max_date}
\alias{get_min_max_date}
\title{Get min and max date from measurement table}
\usage{
get_min_max_date(con, station_id)
}
\arguments{
\item{con}{PqConnection: database connection}

\item{station_id}{integer: station id}
}
\value{
data.frame
}
\description{
Get min and max date from measurement table
}
\examples{
con <- db_con()
get_min_max_date(con, 3)
}
