% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/louroux-package.R
\docType{package}
\name{louroux-package}
\alias{louroux}
\alias{louroux-package}
\title{louroux: Louroux}
\description{
Louroux data processing.
}
\author{
\strong{Maintainer}: Louis Manière \email{louis.maniere@univ-tours.fr}

}
\keyword{internal}
