% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_params.R
\name{params_get_authors}
\alias{params_get_authors}
\title{authors list}
\usage{
params_get_authors(con)
}
\arguments{
\item{con}{PqConnection: database connection}
}
\value{
list
}
\description{
This function returns a list of authors from the database.
}
