% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_params.R
\name{params_get_correction_type}
\alias{params_get_correction_type}
\title{Correction type list}
\usage{
params_get_correction_type(con)
}
\arguments{
\item{con}{PqConnection: database connection}
}
\value{
data.frame
}
\description{
This function returns a list of the type of correction available from the database.
}
