% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_params.R
\name{params_get_device_models}
\alias{params_get_device_models}
\title{Get device models list}
\usage{
params_get_device_models(con)
}
\arguments{
\item{con}{PqConnection: database connection}
}
\value{
vector
}
\description{
Get device models list
}
