% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_params.R
\name{params_get_devices}
\alias{params_get_devices}
\title{Get devices list}
\usage{
params_get_devices(con)
}
\arguments{
\item{con}{PqConnection: database connection}
}
\value{
vector
}
\description{
Get devices list
}
