% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_params.R
\name{params_get_interval}
\alias{params_get_interval}
\title{Get main interval sensor measurement.}
\usage{
params_get_interval(con, sensor_id)
}
\arguments{
\item{con}{PqConnection: database connection}

\item{sensor_id}{integer: sensor id}
}
\value{
list
}
\description{
Get main interval sensor measurement.
}
