% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_params.R
\name{params_get_sensor_id}
\alias{params_get_sensor_id}
\title{Sensor id}
\usage{
params_get_sensor_id(con, station_code, parameter_name)
}
\arguments{
\item{con}{PqConnection: database connection}

\item{station_code}{string: station code}

\item{parameter_name}{string: parameter name}
}
\value{
list
}
\description{
This function returns the sensor id based on the station code and the parameter name.
}
