% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_params.R
\name{params_get_station_id}
\alias{params_get_station_id}
\title{Station id}
\usage{
params_get_station_id(con, station_code)
}
\arguments{
\item{con}{PqConnection: database connection}

\item{station_code}{string: station code}
}
\value{
list
}
\description{
This function returns the station id based on the station code.
}
