% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_params.R
\name{params_get_stations}
\alias{params_get_stations}
\title{stations list}
\usage{
params_get_stations(con)
}
\arguments{
\item{con}{PqConnection: database connection}
}
\value{
vector
}
\description{
This function returns a list of stations from the database.
}
