% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_params.R
\name{params_get_table_fields}
\alias{params_get_table_fields}
\title{Get table fields}
\usage{
params_get_table_fields(table_name, con)
}
\arguments{
\item{table_name}{character: table name}

\item{con}{PqConnection: database connection}
}
\value{
data.frame
}
\description{
Get table fields
}
