% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot.R
\name{plot_add_corr_trace}
\alias{plot_add_corr_trace}
\title{plotly add corr trace.}
\usage{
plot_add_corr_trace(data, y, y_label)
}
\arguments{
\item{data}{data frame containing the selected axis data.}

\item{y}{text metric to be plotted on the y-axis.}

\item{y_label}{text name of the metric plotted.}
}
\value{
list
}
\description{
plotly add corr trace.
}
