% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot.R
\name{plot_add_missing_period}
\alias{plot_add_missing_period}
\title{plotly add missing periods.}
\usage{
plot_add_missing_period(data)
}
\arguments{
\item{data}{data frame containing the selected axis data.}
}
\value{
list
}
\description{
plotly add missing periods.
}
