% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot.R
\name{plot_add_valid_period}
\alias{plot_add_valid_period}
\title{plotly add available periods.}
\usage{
plot_add_valid_period(data)
}
\arguments{
\item{data}{data frame containing the selected axis data.}
}
\value{
list
}
\description{
plotly add available periods.
}
