% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot.R
\name{plot_available_raw}
\alias{plot_available_raw}
\title{plotly add available periods.}
\usage{
plot_available_raw(station_id, date_start, date_end)
}
\arguments{
\item{station_id}{integer station id.}

\item{date_start}{date start date.}

\item{date_end}{date end date.}
}
\value{
plotly graph
}
\description{
plotly add available periods.
}
