% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot.R
\name{plot_dygraph}
\alias{plot_dygraph}
\title{Create a dygraph plot for the given data and parameters.}
\usage{
plot_dygraph(
  data,
  parameter_tocorr_name,
  parameter_add_name,
  y_title,
  y_title_add,
  display_opts
)
}
\arguments{
\item{data}{A list containing the measurement data to be plotted.}

\item{parameter_tocorr_name}{The name of the parameter to be corrected.}

\item{parameter_add_name}{The name of the additional parameter to be plotted (optional).}

\item{y_title}{text the parameter name and unit.}

\item{y_title_add}{text additionnal parameter name and unit.}

\item{display_opts}{A list of boolean values indicating which series to display (raw, corrected, additional).}
}
\value{
A dygraph object that can be rendered in a Shiny app or R Markdown document.
}
\description{
Create a dygraph plot for the given data and parameters.
}
