% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot.R
\name{plot_field_lines}
\alias{plot_field_lines}
\title{Create a list of vertical dashed line annotations for longitudinal profile plots}
\usage{
plot_field_lines(data)
}
\arguments{
\item{data}{A data frame containing the x-coordinates of the vertical lines.}
}
\value{
A list object containing the vertical dashed line annotations.
}
\description{
Create a list of vertical dashed line annotations for longitudinal profile plots
}
