% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot.R
\name{plot_lines}
\alias{plot_lines}
\title{Create a list of vertical dashed line annotations for longitudinal profile plots}
\usage{
plot_lines(data, color = "green")
}
\arguments{
\item{data}{dataframe containing the x-coordinates of the vertical lines.}

\item{color}{text color of the vertical dashed lines (default is "green").}
}
\value{
A list object containing the vertical dashed line annotations.
}
\description{
Create a list of vertical dashed line annotations for longitudinal profile plots
}
