% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot.R
\name{plot_main}
\alias{plot_main}
\title{plot_main}
\usage{
plot_main(data, y, y_title, date_min, date_max)
}
\arguments{
\item{date_min}{minimum date xaxis}

\item{date_max}{maximum date xaxis}
}
\value{
plotly graph.
}
\description{
plot_main
}
