% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot.R
\name{plot_update_main}
\alias{plot_update_main}
\title{Update plotly graph.}
\usage{
plot_update_main(data, y, y_title, date_min, date_max)
}
\arguments{
\item{data}{dataframe sensors dataset.}

\item{y}{text parameter selected.}

\item{y_title}{text parameter selected name.}

\item{date_min}{minimum date xaxis}

\item{date_max}{maximum date xaxis}
}
\value{
list with proxy trace and layout.
}
\description{
Update plotly graph.
}
