% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot.R
\name{plot_vertical_line}
\alias{plot_vertical_line}
\title{Create a vertical dashed line annotation for longitudinal profile plots}
\usage{
plot_vertical_line(x = 0, color = "green")
}
\arguments{
\item{x}{The x-coordinate where the vertical line should be positioned.}

\item{color}{The color of the vertical dashed line (default is "green").}
}
\value{
A list object representing a vertical dashed line annotation.
}
\description{
This function generates a vertical dashed line annotation for longitudinal profile
plots using the 'plotly' package.
}
