% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_compile.R
\name{compile_raw}
\alias{compile_raw}
\title{Compile raw data into database}
\usage{
compile_raw(con, station, parameter, sensor)
}
\arguments{
\item{con}{PqConnection: database connection}

\item{station}{character: station code}

\item{parameter}{character: parameter name}

\item{sensor}{integer: sensor id}
}
\value{
string: message
}
\description{
Compile raw data into database
}
