% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_compile.R
\name{compile_raw}
\alias{compile_raw}
\title{Compile raw data}
\usage{
compile_raw(station, parameter, append = FALSE, col_names = TRUE)
}
\arguments{
\item{station}{string: station name}

\item{parameter}{string: parameter name}

\item{append}{logical: append to existing file}

\item{col_names}{logical: write column names}
}
\value{
string: message
}
\description{
Compile raw data
}
\examples{
compile_raw("BE", "level")
}
