% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_compile.R
\name{compile_raw}
\alias{compile_raw}
\title{Compile raw data into database}
\usage{
compile_raw(con, station, parameter)
}
\arguments{
\item{con}{PqConnection: database connection}

\item{station}{string: station name}

\item{parameter}{string: parameter name}
}
\value{
string: message
}
\description{
Compile raw data into database
}
