% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data.R
\name{data_get_deleted_periods}
\alias{data_get_deleted_periods}
\title{Format deleted period from deleted threshold}
\usage{
data_get_deleted_periods(
  dataframe,
  sensor_id,
  delete_threshold,
  author_id,
  correction_type,
  comment
)
}
\arguments{
\item{dataframe}{data.frame: data frame with ts, value, value_corr and value_edit columns}

\item{sensor_id}{integer: sensor id}

\item{delete_threshold}{numeric: threshold value to consider a value as deleted}

\item{author_id}{integer: author id}

\item{correction_type}{integer: correction type id}

\item{comment}{character: comment}
}
\value{
data.frame
}
\description{
Format deleted period from deleted threshold
}
