% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_db.R
\name{db_apply_edit_with_correction}
\alias{db_apply_edit_with_correction}
\title{Apply measurement edit and correction periods in a single transaction}
\usage{
db_apply_edit_with_correction(con, measurement_edit, correction_period)
}
\arguments{
\item{con}{DBIConnection}

\item{measurement_edit}{data.frame with columns ts, sensor_id, value, value_corr and value_edit (can be NULL if no measurement edit)}

\item{correction_period}{data.frame with columns sensor_id, author_id, ts_start, ts_end, correction_type, value, comment (can be NULL if no correction period)}
}
\value{
character Message indicating the result of the operation
}
\description{
This function applies both measurement edits and correction periods in a single database transaction. If any part of the process fails, the entire transaction is rolled back to maintain data integrity.
}
