% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_db.R
\name{db_get_measurement_raw_corr}
\alias{db_get_measurement_raw_corr}
\title{Get cleaned data with correction applied}
\usage{
db_get_measurement_raw_corr(con, sensor_id, min_date, max_date)
}
\arguments{
\item{con}{PqConnection: database connection}

\item{sensor_id}{integer: sensor id}

\item{min_date}{POSIXct: minimum date}

\item{max_date}{POSIXct: maximum date}
}
\value{
data.frame
}
\description{
This function returns the measurements based on the sensor id and the date range.
Return corrected data if possible, else it return raw ones.
}
