% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_db.R
\name{db_insert_data}
\alias{db_insert_data}
\title{Insert data into a table}
\usage{
db_insert_data(db_conn, table_name, input_values)
}
\arguments{
\item{db_conn}{PqConnection: database connection}

\item{table_name}{character: table name}

\item{input_values}{list: list of input name and values}
}
\value{
list
}
\description{
Insert data into a table
}
