% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_db.R
\name{db_min_max_date}
\alias{db_min_max_date}
\title{Min and Max date}
\usage{
db_min_max_date(con)
}
\arguments{
\item{con}{PqConnection: database connection}
}
\value{
data.frame
}
\description{
This function returns the minimum and maximum date of the measurements.
}
