% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_db.R
\name{db_update_measurement_edit}
\alias{db_update_measurement_edit}
\title{Update measurement table with edited values}
\usage{
db_update_measurement_edit(con, dataframe)
}
\arguments{
\item{con}{DBIConnection}

\item{dataframe}{data.frame with columns ts, sensor_id, value, value_corr and value_edit}
}
\value{
integer Number of rows updated
}
\description{
Update measurement table with edited values
}
