% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_params.R
\name{params_get_parameters}
\alias{params_get_parameters}
\title{parameters list}
\usage{
params_get_parameters(con, station_code)
}
\arguments{
\item{con}{PqConnection: database connection}
}
\value{
list
}
\description{
This function returns a list of parameters from the database.
}
