% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_params.R
\name{params_get_sensor_id}
\alias{params_get_sensor_id}
\title{Sensor id}
\usage{
params_get_sensor_id(con, station_id, parameter_id)
}
\arguments{
\item{con}{PqConnection: database connection}

\item{station_id}{integer: station id}

\item{parameter_id}{integer: parameter id}
}
\value{
list
}
\description{
This function returns the sensor id based on the station and parameter id.
}
