% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot.R
\name{plot_main}
\alias{plot_main}
\title{Plot a plotly graph with selected parameter.}
\usage{
plot_main(data, y, y_title)
}
\arguments{
\item{data}{dataframe sensors dataset.}

\item{y}{text parameter selected.}

\item{y_title}{text parameter selected name.}
}
\value{
plotly graph.
}
\description{
Plot a plotly graph with selected parameter.
}
